Wishlist - ViewModel
